+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix a bug where the EntitySelect component `multiple` option is broken - #47
+
 ## [1.0.0-alpha12] - 2024-04-09
 
 ### Changed